Update bundle-gen to create release-config
Since Hive started using File Based Catalog system on the Red Hat
OpenShift Ecosystem OperatorHub, we are required to upload a
release-config.yaml file along with our bundle, which lists the replaces
version, channel and the template mapping. The presence of this file
with the bundle ensures an automatic PR post merge that will update all
the catalogs.

Note: This file is only useful for RH ecosystem, however we build the
same bundle for both OperatorHubs and the Kubernetes Ecosystem
OperatorHub ignores this file.

Signed-off-by: Suhani <sumehta@redhat.com>