validation-gen naked import
Since validation generator was added to upstream update-codegen.sh and
the corresponding changes were carried in code-generators
kube-codegen.sh (called by our hack/update-codegen.sh on make build),
the script would try to go install validation-gen binary but error out
because some environments do not have GOBIN set.

The fix (https://github.com/kubernetes/kubernetes/pull/132378) was since
introduced and backported to v1.34, but it hasn't been added to v1.33.
Hence, we have a temporary workaround to include a naked import for that
binary, and we can remove this patch when kube is bumped to v1.34.