Update build images to 1.24
This also requires bumping OCP version to 4.20 because 4.19 is not
available with go1.24

skopeo list-tags docker://registry.ci.openshift.org/ocp/builder | grep 1.24
        "rhel-9-golang-1.24-nofips-openshift-4.19",
        "rhel-9-golang-1.24-openshift-4.20",
        "rhel-8-golang-1.24-openshift-4.21",
        "rhel-8-golang-1.24-openshift-4.20",
        "rhel-9-golang-1.24-openshift-4.21",

Drop "Using public images" section from docs

RHEL based images we use in our Dockerfile or Makefile requires a user
to be logged into the registry, so "Using public images" section in the
developing doc suggested an alternative that would work with publically
available images.

However, there isn't a publically available image with golang1.24 on
registry.ci.openshift.org/openshift/release as of now; and, we are yet
to decide on an alternative that would be feasible.

For now, this commit drops that section from the docs.