NO-ISSUE: snyk: Unignore SNYK-GOLANG-GITHUBCOMOPENSHIFTINSTALLERDATA-1070553
https://security.snyk.io/vuln/SNYK-GOLANG-GITHUBCOMOPENSHIFTINSTALLERDATA-1070553

Due to a semver comparison error, snyk was incorrectly flagging the
above issue in our vendor of openshift/installer. We added an ignore
rule via #2611 / 0210417 to keep our `security` CI job green.

In response to support case 00104986, snyk updated their database, so we
can now get rid of this rule.