HIVE-2302, HIVE-2644: Destroyers use metadata.json
An earlier commit ensures that ClusterDeployments have an associated
Secret containing the metadata.json emitted by the installer.

This change adds a new generic destroyer via the (existing) `hiveutil
deprovision` command that consumes this metadata.json to deprovision the
cluster.

This new behavior is the default, but we also include an escape hatch to
run the platform-specific legacy destroyer by setting the following
annotation on the ClusterDeployment:

`hive.openshift.io/legacy-deprovision: "true"`