Add non-resource regex in errorScrub for AWS AccessDenied errors
We have a regex that we use to scrub AWS errors for AccessDenied or
NotAuthorized operations. That regex looks for both the User ARN and the
resource ARN that gives the AccessDenied.

However, when firing a Describe for resources (like a loadbalancer), we
might not receive the resource ARN in the error message, so errorScrub
skipped the scrubing. This can result in thrashing - especially in STS
AssumeRole clusters because the User ARN is returned with a unique
identifier every time.

This commit adds a regex similar to the one we already have, just not
expecting a resource towards the later half of the string.

A unit test that simulates the actual error is also included to prove
the regex works and does not interfere with the regex we already have.