fix(pkg/controller/remoteingress): remove status and prune spec fields of created IngressController resources
Remove the status and prune spec to a strict allowlist of user-settable fields derived from ClusterDeployment.
This prevents overwriting changes made by other controllers or users to the IngressController resources.
For example, it could be desirable to set the httpCompression field on the IngressController. Before this commit
this was not possible because it would have been overwritten.