HiveSLO - update env_name query
On HiveSLO grafana dashboard, a env_name variable is defined to fetch
deployments for each shard. The query for the same is looking for
specific values that can go out-of-sync when a new shard is created.
This commit switches the query to use a regex that can filter out all
the relevant values, thus making it a more dynamic, as long as the
shards are named following the convention. The only shard that doesn't
follow the naming convention is osd-stage-01, which is included
explicitly.

Note that this variable isn't actually used for the dashboard as of now
because the query is disabled. Regardless this change would ensure the
query is same as that of Hive Metrics dashboard where the same variable
is used as "Deployment Environment"