privatelink: pass privatelink enabled status into actuators
Previously, the aws hub actuator was using its own platform logic to
determine if privatelink was enabled on the cluster deployment. This
caused a nil pointer exception when deleting non-aws clusters with
preserve on delete because the spec.platform.aws value was nil. More
importantly, the logic was flawed because it should instead be based on
the cluster deployment's platform.

This change passes the privatelink enabled status from the controller to
each of the actuators on creation. The controller was already using the
cluster deploymnet's platform to determine if privatelink is enabled.
Passing this value into the actuators enables the aws hub actuator to
properly determine if cleanup is required while avoiding the nil pointer
exception altogether.