HIVE-2302: Refactor ConfigureCreds to modify installer metadata
A couple of providers (nutanix, vsphere) need bespoke code to populate
credentials in the metadata.json object for destroying a cluster. In a
prior commit this was being done in the deprovisioner (the new one, that
uses metadata.json directly, per HIVE-2302) after ConfigureCreds.

Since ConfigureCreds is where we (stay with me) configure creds, and is
already platform-specific, it makes more sense to do this work there.
This commit refactors to do so.

Legacy code paths pass in a `nil` metadata object, which is coded to
result in no change from the previous functionality. (In particular,
ConfigureCreds is also used when provisioning, where no metadata object
is present/necessary.)