ACM-26271: Only vivify ClusterMetadata.Platform for fake clusters
For purposes of this work, there are three kinds of clusters:
- IPI, where hive manages the provisioning via installer.
- UPI, where a ClusterInstall implementation provisions the cluster and
  hive just watches and copies over the status.
- Fake, used for testing purposes.

When we implemented HIVE-2302 / #2729, we didn't account for
UPI/ClusterInstall, which we don't really have a good way to test, and
ended up injecting a bug:

Via that effort, we started populating a new Secret containing
metadata.json produced by the installer. For legacy clusters (those that
existed before upgrading to a version with this feature) we need to
retrofit that Secret based on the ClusterMetadata (among other things),
which was previously how we saved off the metadata.json. For IPI, that
ClusterMetadata always had a Platform section. The changes we had to
make for fake clusters saw us spoofing a very sparse metadata.json and
then populating it later. That process relied on the existence of the
CD.Spec.ClusterMetadata.Platform section, so we were creating it for the
sake of that fake cluster path. However, it turns out that
ClusterInstall subclasses don't (and don't need to) populate the
ClusterMetadata.Platform section. Since we blindly copy the
ClusterMetadata into the ClusterDeployment, we could end up with that
Platform section being absent when we come to retrofit the
metadata.json. We would then hit the path designed for fake clusters
where we would create that section (empty). No problem, right? Except
that we have a validating admission webhook that forbids making changes
to the ClusterMetadata section, and that new, empty Platform field was
flagged as such a change, and bounced by the webhook.

Phew.

So with this change, we condition populating that Platform section
explicitly on fake clusters only.