HIVE-3017: Trim whitespace from vsphere credentials
It seems VCF9 is not as tolerant of whitespace (notably trailing `\n`)
in credentials as vcenter 8. While it should be the user's
responsibility to make sure their creds are "clean", we'll trim
whitespace in two out of three paths (see below) to help out:
- preflight credentials check
- destroyer

Note that we're *not* breaking open the install-config.yaml to "fix"
padded credentials there for the installer path.